+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [5.5.0] - 2026-08-17
+
+### Added
+- Added `CreateAssetMenuMarkerAttribute` for declaring reusable ScriptableObject creation categories through marker interfaces.
+- Added the `Assets/Create/Project` menu with nested paths and support for ScriptableObjects implementing multiple marker interfaces.
+
 ## [5.4.2] - 2026-08-16
 
 ### Changed